project/somaz-bigquery/cloud-functions/matic-value-to-sheet/requirements.txt
gspread==5.4.0
google-auth==2.6.6
requests==2.27.1
